test(server): split ha serving mode coverage
