+++ scripts/ci_backend_test_manifest.json
@@ -449,6 +449,7 @@
         "server::tests::linuxdo_oauth_and_admin_keys::manual_",
         "server::tests::upstream_support_and_manual_jobs::manual_",
         "server::tests::alerts_and_ha::ha_",
+        "server::tests::alerts_and_ha_serving_modes::ha_",
         "server::tests::system_settings_and_forward_proxy::startup_",
         "server::tests::admin_token_filters_and_maintenance::maintenance_",
         "server::tests::admin_logs_and_summary::summary_",
